Tilemap: Offset generated sprites by 1/2 W,H
Needed due to how I'm generating the sprite at the corner between 4
input terrain values.  Will provide a graphic to demonstrate this once
I've cleaned this up.

Due to this, the input file extents must add an extra row and column
around all sides of the map to be displayed